Add "useicon" option, which uses inventory icons
This was mostly done to have it display properly for the ZM66(had
incorrect size) and rations in UaS(it just didn't display at all,
because it's CurState didn't have a sprite).

As the name suggests, if stated, will use inventory icon instead.
Seems to be a bit janky, but it works for now.