Fix player items not glowing after a level
Holy shit, this was a pain in the ass to fix.
It seems like it would be easy to fix, but nothing ever comes easy now
does it?

A few things that I've learned from this experience:
* Player doesn't spawn in when moving to the next level(?)
* The Inventory is weird as shit
* Player's Inventory doesn't get loaded at the same time as the player
spawns

Also, removed some unneccessary Destroy() calls, since EventHandler only
lasts for a level. (found that out the hard way)